requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.8.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-dotenv==1.0.0
redis==5.0.1
